+++ src/types/frame.ts
@@ -1,5 +1,3 @@
-import type { FarcasterNetwork } from '@farcaster/hub-web';
-
 export type Index = 1 | 2 | 3 | 4;
 
 export enum ActionType {
@@ -44,23 +42,4 @@ export interface Frame {
 
 export interface LinkDigested {
     frame: Frame;
-}
-
-export interface FrameSignaturePacket {
-    untrustedData: {
-        fid: number;
-        url: string;
-        messageHash: string;
-        timestamp: number;
-        network: FarcasterNetwork;
-        buttonIndex: number;
-        inputText?: string;
-        castId: {
-            fid: number;
-            hash: string;
-        };
-    };
-    trustedData: {
-        messageBytes: string;
-    };
 }
